Stavki3/requirements.txt
# =============================================================================
# BETTING ASSISTANT V2 — ALL PHASES (V2.0 → V2.4)
# pip install -r requirements.txt
# =============================================================================

# ── Core ──
python-dotenv==1.0.1
aiohttp==3.9.5

# ── Telegram Bot ──
aiogram==3.13.1

# ── Database ──
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
redis[hiredis]==5.2.1

# ── ML / Data Science (V2.1) ──
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.2
catboost==1.2.7

# ── Football Data (V2.2) ──
soccerdata==1.8.2        # FBRef, Understat, ClubElo scrapers
penaltyblog==1.6.2       # Dixon-Coles optimized (Cython)
beautifulsoup4==4.12.3   # HTML parsing fallback

# ── Deep Learning (V2.3, optional) ──
# torch>=2.0.0           # Uncomment for LSTM line predictor
# torchvision>=0.15.0

# ── Web Dashboard (V2.4) ──
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==13.1

# ── Rating Systems ──
elote==1.1.0

# ── Utilities ──
httpx==0.27.2
schedule==1.2.2

# ── Dev / Testing ──
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.4
